Move ConflictReason to semantic module of bee-message
