+++ bee-tangle/CHANGELOG.md
@@ -24,11 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use `packable` instead of `bee-common::packable` for serialization and deserialization.
-- Variants of `ConflictReason`;
 
 ### Removed
 
-- `ConflictReason::InvalidDustAllowance`;
+- `ConflictReason` moved to `bee-message`;
 
 ## 0.3.0 - 2022-XX-XX
