+++ bee-test/src/rand/metadata.rs
@@ -5,10 +5,10 @@ use crate::rand::{
     message::rand_message_id, milestone::rand_milestone_index, number::rand_number, option::rand_option,
 };
 
+use bee_message::semantic::ConflictReason;
 use bee_tangle::{
     flags::Flags,
     metadata::{IndexId, MessageMetadata},
-    ConflictReason,
 };
 
 /// Generates a random conflict reason.
